expression can be divided in subexpressions

+++ backend/internal/agent/response/response.go
@@ -1,7 +1,7 @@
 package response
 
 type Response struct {
-	Id            int    `json:"id"`
-	Subexpression string `json:"subexpression"`
-	Timeout       int64  `json:"timeout"`
+	Id            int     `json:"id"`
+	Subexpression string  `json:"subexpression"`
+	Timeout       float64 `json:"timeout"`
 }
